Major refactoring and redesign of appliaction. Fleet code seperated into library. Remoted remoting code, everything is WCF. There are three applications for using image transfer: 1. Fleet Daemon. A command line application to boradcast the service. 2. Fleet.Test. The command line application, used to send files from a remote machine. A GUI may be built. 3. FleetViewer. A  WInforms based application to open the received images sent to Fleet Daemon.
FleetDaemon and FleetViewer communicate via WCF pipes.
FleetDaemon and Fleet.Test communication via WCF TCP sockets.

I can confirm that this works between two windows 10 machines on my local network. Not confirmed if this works on Mono.
Fleet service name bradcast is not implemented, all broadcasts will use the service name Fleet. This is not a huge issue, but will be rectified in a later update.